fix(geoip): make cache thread local (#1323)
This was initially submitted as #1314. Then it was reverted in #1321,
because iOS does not support `thread_local` (#1319).

However, [research](https://stackoverflow.com/a/29929949) suggests that
it is actually supported for iOS >= 9.0.

As of today, the market share of iOS >= 9.0 is 98.58% (data gathered
from [apteligent](https://data.apteligent.com/ios/) and subject to change
in the future, of course).

It seems to me it makes a lot of sense to remove support for iOS < 9 and
have the `thread_local` feature. If we _really_ need iOS 8 support, I can
write some iOS 8 specific code using pthreads.

I really prefer a solution based on `thread_local` storage rather than
one with a global cache for which we should maintain locking and for which
we should have destruction from possibly multiple threads.

Compared to #1314, this diff contains the following changes:

- removed support for iOS < 9

Closes #1319, #1209.