ssl: refactor for readability, fix thread safety (#1313)
* ssl: refactor for readability, fix thread safety

- Move all SSL code into the same file, to increase readability and
  to simplify reasoning about our SSL usage.

- Do not split implementation between classes and functions, rather
  use template classes for testability. This should increase readability
  and be closer to what a typical C++ programmer would expect.

- Write an explicit SSL cache rather than having an implicit one, which
  also increases the readability of SSL code.

- Use thread local storage for SSL cache, so to solve #1296.

- Add doxygen style comments.

Further details on specific implementation choices are provided in
comments added in the new code. Further comments in the PR.

* Rewrite test/net/libssl.cpp

* Make GCC happy and properly free `SSL *`

Apparently clang is more forgiving with respect to a parent class calling
a protected constructor of a child class. No big deal.

Regarding `SSL *`, ownership is transferred to the caller, so we need to
properly dispose of them when done, or Valgrind will complain.

* fix(test/nettests): wait for background threads

We need to do so because now we allocate thread local storage
for SSL_CTX. Hence, if we don't wait for the thread to exit, we
have valgrind errors because such storage is not released.

The technique used in here is similar to the one that was also
employed for the orchestration.

Basically, we wait for the background tasks queue to be empty
in the ghetto way, i.e. busy looping.

Yet, the busy loop is short, because for the most time the code
will be blocked in run() and the busy loop is only to ensure that
we give enough time to background threads to finish.

* fix(net/libssl): better library initialization

Rather than constructing an object so to have library initialization
as a side effect, use explicit initialization function.

I was worried that a compiler could optimize away the object
construction because the object is not used, thus eliminating
also the side effect. Using a `void(void)` function seems to
me more safe, because it is obvious that you call it for a
purpose (the side effect), and so there is no grounds for a
compiler to optimize it away.

* remove unused headers

* fix: Valgrind thread-local-storage errors

Second attempt. Add more delay after we know that all the
background threads have exited.

Should work, or I have not understood the real issue.

* More coverage