cleanup(reactor): proactor-friendly socket poll funcs
In the event we would ever want to switch to a proactor based
backend like, for example, asio, be prepared by having a reactor
API that is compatible with that.

Costs basically nothing (a few lines or refactoring) but possibly
enables us to do more experimentation in the future.

While there, remove unused code from c/l/reactor.hpp.