Fix for systems that don't support thread_local
There is empirical evidence that MK v0.7.5 crashes on 32 bit iOS systems
because thread_local isn't working properly.

See #1397.

This diff is an attempt to fix the issue in a very light way. Basically,
retain the thread local semantic, but allocate a cache very time if
thread_local is known to misbehave.

This fix makes the code less efficient for those devices where thread_local
is not working, of course. In fact, basically, for those devices any kind of
caching is disabled.

Yet, given that this seems to affect a minority of our users, I've decided
to go for an as-simple-as-possible fix that prevents the crash.

(Also, `armv7` and `armv7s` are going to disappear anyway in the near future,
because Apple is now mostly focusing on `arm64`.)

While there I've refactored `libssl.hpp`'s cache factory to return a shared
pointer rather than a reference, in the hope that, if the shared pointer
is not initialized, this will lead to an exception and not a crash.