Rename raw_ptr as unique_ptr
After reading excerpts of book on C++ recommended by @hellais, I did
convince myself that it's better not to implicitly cast a smart pointer
to its underlying pointer type, but it's better to have an explicit
method, as `get()` for doing that.

After this change, RawPtr now really looks like a wrapper for a unique
pointer, hence just rename it for clarity and propagate changes.