feature(build): script to update .gitignore on demand
Much simpler than before because now we have per dir .gitignore as
opposed to having a single master .gitignore file.

I would say we have not lose any functionality since before.

The main difference is that now one needs to call the script to
update the .gitignore files whereas before it was automatic.

But we want to be able to build on Windows withouting invoking any
script, thus we want to make some things less automatic.