refactor: remove private headers
Private headers below `include` were looking more tidy but at the end of
the day it seems simpler to have such headers as close as possible to
implementation files, because then it's more natural to look for files
and to make sure they are in sync, etc.

While there, make sure the project name is correctly spelled. This was
done by using a one-off script that I did write.