fix(error): define virtual destructor and simplify
The bug we're trying to fix here is that, since Error derives from
std::exception, its destructor should be virtual.

The proper way of doing that would probably be that of defining a
virtual destructor for every error class that we have.

Adding an out-of-line destructor for each error class would take
time and/or being hackish (by augmenting the existing macro?).

Adding an in-line virtual destructor is bad practice (and indeed
clang warns about that; see `-Wweak-vtables`).

However, it seems we really don't need all these classes.

They were introduced when we were thinking about throwing errors
and then catching them using different catch clauses.

But, during #1235, #1262, and #1447, it occurred to me repeatedly
that we should only use exceptions as fatal errors because to have
exceptions and async code in the same plate is too tricky.

So, if that is the intent, we can get away with a single Error
class -- with proper virtual destructor, of course.

At least in the library.

Then, in the regress tests, we have a few instances where we do
actually check that an exception is of the specified type.

But, it also turn out that doing that is not exactly needed for
our code testing goal. Which is that of making sure that the code
throws `$whatever` in case something unexpected happens.

If that is the end, we can safely just check if the code throws
something and then rely on code coverage to make sure that all
the possible fatal error paths throw.

After all, we already have cases where different error paths lead
to the same exception being thrown, so we already cannot use the
exception type alone to tell whether we cover all paths.

While there, force the single-argument constructor to be called
explicitly, as that is considered good practice.

While there, force moving in the child error.