fix(error_or): make construction unambiguous
This diff is a response to @joelanders observation that sometimes
making an ErrorOr was leading to ambiguity.

The way in which I implemented it was that I removed all the
partial constructors, so now you always implement returning
ErrorOr like when returning a tuple. Which is more or less the
thing we're doing anyway. And also is more readable.