feature(third_party): enable Windows builds
This diff introduces build system changes that allow one to build
all the default dependencies for Win32 and Win64 targets.

Tested with Visual Studio Community 2017 and CMake 3.10.0.

Will publish the binaries created in here in the dependency
repositories quite soon.

After this commit, the diffs to be applied after updating the
repository start become dangerously stale. I'll fix that.