cleanup(build): don't save vendor patches
It is so bureaucratic to have them around when developing to
improve our dependencies. Considering that one can always get
the diff between a pristine import and subsequent changes,
since everything is now committed, there is no point in keeping
the patches around. Actually, they can lead to mistakes as one
can modify a dependency without saving the patch and this would
later lead to annoyances. Better to dispose of them.