cleanup(include): stop throwing Error (#1545)
* refactor(error): explicit ctor, move semantic

This diff introduces a set of non-breaking correctness-related
changes to the Error class:

- explicit single-argument constructor

- enforce move semantic when adding child errors

Extracted from #1531. This should not be controversial and we can
perhaps see to merge it while deciding how to further refactor the
Error class according to the discussion in #1531.

* cleanup(include): stop throwing Error

As discussed in #1531, one of the first steps to start simplifying
`Error` is to make sure we are not going to throw `Error`.

Rather, change the code to throw `std::runtime_error`.

The specific reason why I want to avoid throwing `Error` is that I am
worried that that will break assumptions made by code that was using
`catch (const std::exception &)` to _also_ catch for `Error`.

By never throwing `Error`, we can make this issue disappear.

Then we can move forward by detaching `Error` and `std::exception`.

This diff does not address all the cases where we `throw Error` rather
it starts addressing some of these. More diffs/PR will follow.

I am following this process to keep the workload low for reviewers
and to make sure that I can wrap up work at end of day.