fix(build): yakkety is no more, find replacements (#1552)
* fix(build): use artful instead of yakkety

Yakkety (which worked very well for us) is unfortunately not
available anymore, so try with artful instead. I have also tried
with xenial, but [did see too many weird Valgrind errors](
https://travis-ci.org/measurement-kit/measurement-kit/builds/332279102).

* fix(valgrind-clang): repair build with artful

As explained in the XXX comment, libc++ is broken in artful and
expects xlocale.h to be present on the system. Repair by creating
an empty xlocale.h file such that compilation doesn't fail.

* fix(docker/coveralls): use zesty instead of artful

Dunno why artful doesn't work. Coveralls Python submitter for C++ seems
to be broken. Easier to downgrade to zesty than to investigate.

* fix(docker/coveralls): fix previous by using xenial

From-the-what-a-pain-dept. Anyway, I went through coveralls
history and it seems that with xenial it was working:

  https://travis-ci.org/measurement-kit/measurement-kit/builds/332279102

So, use xenial.