fix(measurement_kit): better check for optreset (#1559)
Triggered by @rschulman issues with compiling with Alpine, using the
musl C library. Fix by checking for optreset, which is supported by and
exposed by [musl getopt.h](https://github.com/runtimejs/musl-libc/commit/030e52639248ac8417a4934298caa78c21a228d1).

I am still being a bit of a slacker in that, if `optreset` is not
found, I check whether we're in GNU libc and, otherwise, I say there
is no way to reset the options.

Things might probably be improved with a better (but more painful
to write) macro in m4 for `./configure`.

Since in most cases we'll be using a BSD derived or compatible getopt()
anyway (macOS, Windows, musl), I think this is good enough.

To implement this fix I also considered switching to [clara](
https://github.com/catchorg/Clara) for command line processing,
but it seems it does not support subcommands yet. I will keep
an eye on it, as it would be nice to do the switch.

If this is confirmed to work, it will need forward porting to master.