fix(engine.cpp): remove empty string variables
The Settings and Scalar classes are confused by empty string
variables and treat them as parse error, as @hellais found when
working on Measurement Kit in Go.

To properly fix these classes seems a bit complex. I think it
would be smarter to move forward with this new API, make all
headers except the new API ones private, and then proceed with
very heavy fixing and refactoring on the private headers.

I have tested that this workaround is enough to avoid issues
when working with Measurement Kit in Go.