fix(t/f/scrub.txt): use geoip.ubuntu.com
This is what we use to get the IP address in the first place.

I am not sure this is as powerful as an integration test where we
control the input, because now we have both the first and the
second step of this check tied to the same service.