Fix tests assuming an invalid DNS server would never reply (#1607)
Two tests had this assumption. It's bogus. Especially from my house
where I have Vodafone that always replies to DNS.

Having some tests failing when I run tests locally is so bad.

Fix these instances by setting an unreasonable timeout so that we
basically verify that we deal with the failure.

(Rather than failing because we receive no reply with an invalid
DNS server, now we fail for the super-low timeout.)