Make MK compile on MSYS2 to enable winOONI (#1606)
* Make MK compile on MSYS2 to enable winOONI

This should be enough to have OONI working on Windows. One of the
advantages of using [MSYS2](http://www.msys2.org/) (and hence mingw-w64,
a MINGW fork with enhancements, compiled with POSIX threading model) is
that golang uses the same toolchain, so we easily get go binaries. Such
binaries do however depend on `MSCVCRT.DLL` on which I'll expand later.

Since mingw-w64 is a cross compiler (from GNU/* to Windows) and since it's
available on brew, in theory one may think that it would be possible to
cross compile directly from macOS. Unfortunately, it is not that simple:
The mingw-w64 that you get via brew on macOS [lacks full threading
support](https://github.com/Homebrew/homebrew-core/issues/21706). We
may be able to improve the situation in the future by [using a
replacement](https://github.com/meganz/mingw-std-threads), or we can
recompile the toolchain (which may be a bit painful). Hence, we'll
prepare the MK binaries from Windows and MSYS for now &mdash; no macOS!

Regarding `MSCVCRT.DLL`, the main issue is that each version
of Windows ships with its own `MSCVCRT.DLL`. The only guarantee
provided by Microsoft is that such DLL is ABI compatible with the
subset of such DLL that was provided along with Visual C++ 6.0
(1998). [Thus, applications using that DLL are safe as long as
they limit themselves to the API and ABI that was available in
1998](https://sourceforge.net/p/mingw-w64/wiki2/The%20case%20against%20msvcrt.dll/).

For the records, until Visual Studio 6.0 the practice was to install
MSVCRT.DLL in the system folder. But that was causing issues since
different apps ended up placing there API/ABI incompatible versions
of the library. Therefore, after Visual Studio 6.0 microsoft started
enforcing developers to ship their C runtime library as part of their app,
and such libraries were conveniently versioned. (Another option would
instead be statically linking the app with the C/C++ runtime.) However,
starting from 2015, Microsoft started pusing for a [Universal C Runtime](
https://github.com/MicrosoftDocs/cpp-docs/blob/master/docs/porting/upgrade-your-code-to-the-universal-crt.md).

This Universal C Runtime (Universal CRT or UCRT) is basically a common
version of the runtime that is shipped by default on Windows 10 and
that is automatically installed on older supported systems. As such,
I have a sense that compiling with the UCRT (i.e. using `cl.exe /MD`)
is probably one of the most sensible ways of producing a DLL for the
systems where Windows updates are regularly run.

Creating a Golang binary linking with a URCT-based DLL is something that
actually works and, from some experimental branch, I did this and had it
working on Windows 10. It failed to work on Wine (where it segfaulted),
tho, but I suspect this may be an issue with the UCRT being missing in
the Wine environment. That remains to be investigated.

So, because of unclear issues with the UCRT-based DLL in a very non
standard environment and because the diff to port MK to mingw-w64 was
shorter than the one to port it to MSVC, I decided to finalize the work
on mingw-w64 first. Hence this diff / pull request.

I must say that the compiler has been helpful to spot the places where
MK code was not conforming to the Visual Studio 6.0 subset. As you see
in the diff, in fact, there are several places where I needed to avoid
using printf(), and used std::stringstream instead, because the flags I
was passing to printf() were not defined in 1998. I think there is no way
(for me, at the moment) to say _for sure_ that all the possible issues
with old libraries on old systems have been taken care of. Hovever, for
future reference, I'll note that there is actually a mingw-w64-specific
macro, __MSRVCRT_VERSION__, used by mingw-w64 to select the feature set to
use.  Since by default thay macro is set to a very old value, I will for
now make the reasonable assumption that mingw-w64 only exposes the safe
subset, [unless specifically asked to use a more recent version of the API
and/or to use a more recent DLL](https://stackoverflow.com/a/15865763).

To conclude, this is a first step toward providing Windows binaries. It
should be enough for making a winOONI PoC. Things to keep in mind:

1. we should probably force static linking of libgcc and libstdc++ for
   the binaries so that they work on any Windows

2. we should probably refactor the code to initialize Winsockets so that
   it's a C API exported by MK (and I also wonder whether we need to
   initialize Winsockets from Golang or not &mdash; to be investigated)

3. we should refactor logging, and more generally the library, to use as
   much C++11 as possible, since C++ is often compiled inline so we do
   end up depending on a smaller platform API surface

4. it would be smart to limit usage of threading to a very narrow subset
   of the library and have most of the code assuming single threaded,
   such that working around cross compiling from macOS is easier because
   it entails reasoning and refactoring only a small portion of code

5. I have seen issues with background and asynchronous threads as they
   are used in `measurement_kit.exe` (i.e. segfaults as seen inside of
   `gdb.exe`) but I did not see any issue in code using the new API thus
   I am going to assume we can postpone dealing with this issue since OONI
   uses the new API and I am going to deprecate the old API quite soon

6. we're using libstdc++ whose `<regex>` implementation has given us
   segfaults in the past (see #1276 and #1496). It remains to be seen
   whether this is the case also here (segfaults were with older GNU
   libstdc++ releases, so may this has been fixed now)

I hope you enjoyed reading my commit message :).

Unless there are objections, I'm going to merge this in ~7 days.

* Miscellaneous fixes

- `uname -o` is non portable

- repair build on Unix

- fix all AppleClang warnings