Remove the now-clearly-legacy build system
With the new build system, I was able to:

- compile for Windows on MSYS2

- cross compile for iOS

- cross compile for Android

The resulting binaries seem to work pretty much everywhere.

Check <https://github.com/measurement-kit/script-build-unix>.

While there, remove the never-finished new build system from which
the above mentioned repository is derived.

I know that this commit will break the MINGW build on AppVeyor. I am
going to write a fix for that as part of a subsequent commit.