Run tests with MSVC at AppVeyor (#1623)
* Start easily, just rename the files

* oops

* Start introducing more abstraction

* Try to build for MSVC - will probably fail

* Try with a path

* More fixes to appveyor build

* Fix some more doofus in appveyor.yml

* Attempt to download binaries

* Doofus, doofus everywhere

* Do not just install but also _use_ deps

* Download deps using CMake

* This should repair MSVC _building_

* Attempt fixing the CA bundle for Windows

* msvc.bat: let ctest.exe output on failure

* Reach point where only two MSVC tests fail

The first test that fails is test/report/report which fails for
still unknown reasons. The second test that fails is the
test/http/request one, where the reason for failure is that
we do not correctly gather the error that occurred on Windows,
using `errno` rather than `WSAGetLastError()`.

* Attempt to comment out failing test

The question is: will it fail _also_ at AppVeyor?

* Fix test/common/report MSVC test failure

At this point there is one last test to repair!