Better version of 35825408f1766a4e649ae538954843315bbcfcdc
Apparently making changes to autogenerated code is as dangerous
as wait a few days, forget it's autogenerated and edit.

Luckily there is an integration test for this.

I am not going to tag a release because this change will have
no impact on what is built for Android, iOS, etc.