test/mlabns: s/trn/iad/ to repair the travis build
For not totally clear reasons, mlabns tests that use `trn` are
failing. I have been informed that there are issues with the `trn`
pod by mlab people, hence, as a first step trying to understand
what is going on, I am switching `trn` to `iad`.

Since building this commit in a separate branch led to repair
the build and since we know there are issues with the TRN
we probably should just use IAD and move on.