Remove measurement_kit.podspec
We now use a separate repository to build for iOS. It seems that by
default CocoaPods does not show messages that we are printing in our
measurement_kit.podspec pointing to the other repository.

As such, the user would only see a failure. If such is the case, it
seems then better to just remove this, now historical file.