Replace GeoIP with mk{geoip,mmdb} that use libmaxmindb
Among other things, this code removes the regular expression
approach for extracting the probe IP with another based on
ctypes and characters range checking. However, the latter is
very close in meaning to the regexp based one except for
the fact that it does not check the maximum length of the
IP address string. This is a bug that should be addressed by
the measurement-kit/iplookup implementation before we move
forward with merging this pull request.