Remove the builtin CA bundle
Now that we're starting rolling out cURL based code, I see it
as dangerous that sometimes we could use a builtin CA bundle
(i.e. when not using cURL) and sometimes we cannot.

It's better to start enforcing using an external CA bundle
for all websites always and starting from now.