Start removing CMake build
A single build system is already enough to maintain. All the
new building blocks like mkcurl are compiled on MSVC during CI
tests already. In this repo, let's focus on building for the
systems where it matters the most for OONI. So, no MSVC, hence
no need to have a CMake based build.