t/o/utils.cpp: repair regress test to use real binary data
The previous sequence was incorrect in that, strictly speaking, it was
valid UTF-8 even though it did not contain printable characters.

In fact, the new code using mkdata correctly returned the expected
JSON string representation that we also have in Python.

Fix the regress test by using invalid UTF-8 so to force using base64.