Replace previous JSON errors with JsonProcessingError
With the new code, we cannot distinguish easily between the kind
of JSON errors and honestly, having the logs, it does not even
matter much. You can always check the logs and understand.

Repairs regress tests. The last bit of action to be done is to
inspect all the catch clauses and make sure we are not catching
by mistake any error previously thrown by nlohmann::json.