Honour cookies during HTTP redirects
Headers are actually a list, not a map. This diff changes the code
and makes headers a list. There are parts of the code to which this
change cannot be propagated easily because that would break the
data format. For now, I've opted to avoid breaking it.

Still, this raises the question of how we're dealing with headers
and in particular with `Set-Cookie`, which disappears in the final
report sent to OONI. We should discuss this.

Closes #1547.