fix logging: make sure we base64 non-UTF-8 bodies
We recently enabled LOG_DEBUG2 when running OONI mobile in debug
mode. However, we also recently changed the code to emit logs using
the JSON format. As a consequence, we need to properly quote the
emitted body, when it is not UTF-8, otherwise the JSON library will
crash when trying to serialise the log event.

FWIW, the new HTTP code (https://github.com/measurement-kit/mkcurl)
has an API that explicitly forces you to reason about the body being
possibly binary. When we wrote the code that currently is inside of
MK, we did not consider this risk of logging a binary body because we
were just writing strings on a C++ stream. When we switched to a model
where log events were JSON, we did not consider this issue.

Closes #1699.