web_connectivity: revert testing also http with https URL
Today, I discussed this issue further with @lorenzoPrimi and it
seems we're both not convinced that this is a desirable functionality
to have in OONI mobile v2.0, because, basically, the app is really
not designed to work like this.

The most compelling argument that @lorenzoPrimi brought, in my
opinion, is how to correctly estimate the elapsed time in light
of the fact that more URLs are tested.

While I can see many hackish ways to implement this, I also believe
it's not the job of MK to perform this sort of magic without the apps
knowing it. I'd rather implement in MK a function that takes a list
of URL in input and emits a normalised/expanded list in output.

This is the reason why I'm removing the existing code, rather than
just disabling it. We should discuss more and come out with a proper
solution, which is probably not changing the URLs like we do now.

Related to #1685.