s/l/d/query.cpp: don't assume CNAME and A{,AAA} are together
Apparently, the getaddrinfo()s we've been using so far coalesce
into the same record the CNAME and the A{,AAA} records.

Thus, we never run into the issue described in #1727. The resolver
in Alpine is musl and behaves differently. This behaviour is also
compliant with the standard. So, make MK more smart and ensure that
we avoid attempting to connect to an empty address.

Closes #1727.

Needs backporting to the v0.9.2 branch.