Completely removed global reactor and logger
As mentioned in the commit message of 4f3447ce, it is perhaps
best to completely remove all globals. Otherwise, we would have
mixed globals and non globals in the unit tests.

So, this is the diff that is mentioned in the commit message
of b7f36d25. I already used it to got rid of globals in the
library. I hoped to avoid the burden of changing all the tests
but it seems trying to avoid this was generating headaches.