Vendor also mkbouncer mkcollector mkreport
They are vendored as public, internal headers. Therefore, it seems legit
to write code using such headers for OONI, without committing for long
term API stability with anyone. I believe #1666 should be updated to add
this consideration as an additional benefit for the strategy based on
small building blocks. (TBH, what remains to be clarified is whether to
have a single ObjC/Java code layer for all, or one for OONI and one
for everyone else, where OONI's code layer bears with it more goodies
but also more weigth, e.g., because we include MMDB databases.)