Make the logger API completely private
The last consumer of this API, the orchestration code, has been
made private. As such, also the logger can be private now.

This enables more aggressive refactoring and simplification of
the logger code. The only aspect that continues to be public
is the set of macros used to declare logging levels.