Update to most recent mk{collector,report}
This brings with it the need to delete the wrappers for resubmitting
reports previously written in #1748.

As mentioned in measurement-kit/mkcollector#10 and in #1743, yet, this
deletion is conducive to the import of code that allows us to have
similar functionality with less code and complexity.

I'll do that in the next commit.