Add FFI API for report
This API should allow writing third party tests implemented
in ObjectiveC or Java (or having such an API).

This API reintroduces usage of mk::report::Report but the class
that was previously named like this has been renamed in the previous
commit, so we'll be fine from all points of view.