Rename Report as ReportLegacy
I am about to integrate measurement-kit/mkreport, which also uses the
mk::report::Report name. I think it's not so good to have two classes
with the same name. (Not because that will make the code misbehave --
it will work as long as we use the two classes in different translation
units and they're not exposed -- but rather because it may complicate
reading the code.) So, let's rename the legacy report to avoid conflict.