runnable: make sure we always have a failure key
Closes #1758.

I played a bit with DASH settings without being able to cause
the test to produce a failure-less result.

It does not matter much. This diff makes the code more robust
anyway, because we'll make sure we have a failure.

Also, in the original bug report, that was the case. So, it
seems we should christen MK v0.9.5.