Tag v0.9.6 to have a working stable release
It was pointed out that it's perhaps better that the latest
relase is stable. In all regards, this is v0.9.4 because
of the issue explained in #1761. So, people should actually
use v0.9.4, but this v0.9.6 tag exists such that we have a
stable release as the tip of the stable branch.

Suggested by @evfirerob.

Individual commits:

- release 0.9.6: step 1: update version number

- release 0.9.6: step 2: update headers

Using a PR with a merge commit to clearly mark this is as
a point in the stable history when something happens.