Remove the multi NDT test
With the upcoming M-Lab platform upgrade, most of if not all of the
neubot slice will not be ported to the new platform.

This includes botticelli, the multi stream NDT server that I
wrote some years ago to work with MK's multi stream NDT.

I want to the next release of MK to be a reliable basis for OONI, and
possibly for others, for some months. Therefore, it seems to me that
the best moment for removing multi NDT is now.

Thus, if the platform upgrade happens when MK v0.10.0 is the stable
release of MK, we're not hurting anyone because we've removed the
only test that by default used a weird feature of the neubot slice.

The DASH server will be forward ported to the new platform since
removing it will be a problem for the current OONI, but it's clear
that also DASH needs to be eventually considered for removal or
upgrade to something that supports HTTPS.

Also, forward porting a DASH server to the new platform is just
a few lines of golang. Also, in the [goodbye neubot](
https://neubot.org/2019/01/retiring-neubot-client.html) blog
post I already committed to do that.