collector: write code to resubmit report
It turns out doing it in the apps it too much toil. So write the
code to do that in Measurement Kit.

This will be merged in master, such that the next version of
MK will have it. I'll also vendor a copy of this code inside of
mkall such that the time to release is shorter than having to
perform a full rebuild of MK for all platforms.