runnable: defend against measurement being non dumpable
Related to #1823. It does not really solve the issue, rather it
just prevents Measurement Kit from throwing.

It's still not super pleasant as a remediation, because we
are replacing the whole measurement with a single key telling
us that there's something undumpable.

Yet, in non-anticipated cases where this could happen it is
probably more actionable to us that the app crashing.

The proper fix is to make HIRL more robust.