Remove files unnecessary for the MK build
They are private APIs. So, I feel entitled to remove them in any
moment without causing any API number changes.

Of course, these files are not useless. They are useful in mkall-ios
and android-libs to generate bindings for APIs used by OONI.

Yet, vendoring them in such repositories is useful to increase our
velocity, because a change in any such file does not require to pass
through a full recompilation of MK. What's more, they are the place
where we're applying most innovation currently. This means that such
faster cycles are not theoretical but actually a pain point.

Take for example the recent issue with reuploading. It could have
been solved more quickly with such arrangement.

Another aspect to discuss here is what is the implication in terms
of API wrapping. However, it seems to me, by inspection, that we
just need to change the macros we're using to directly access fields
of C++ objects rather than using the corresponding C wrappers.

Therefore, another byproduct is that we're reducing the amount of
code (and especially boilerplate) that we need to maintain.

Related to: https://github.com/measurement-kit/measurement-kit/issues/1867

Closes: https://github.com/measurement-kit/measurement-kit/issues/1868