ffi: guarantee UTF-7 serialisation
This is perhaps an even more robust strategy to guarantee that Android
is not going to balk at the strings that we emit.

We make sure that we only emit UTF-7 chars. We additionally make sure
that we're not ever going to emit `U+0000` into a serialization.

Obtain that with code and testing.

Still, I want also to keep the guts of engine::Task available as a
private header, just in case I need to perform additional magic.

Closes #1877.