Address feedback by @lorenzoPrimi
Include the 😇 emoji which is U+1F607, thus greater than the maximum
character allowed by modified UTF-8, i.e. U+FFFF.

This is, for example, explained in [this article on Oracle's website](
https://www.oracle.com/technetwork/articles/java/supplementary-142654.html):

> This article describes how supplementary characters are supported
> in the Java platform. Supplementary characters are characters in
> the Unicode standard whose code points are above U+FFFF, and which
> therefore cannot be described as single 16-bit entities such as the
> char data type in the Java programming language.

By including the emoji in the test case, we also make sure that we
properly deal with supplementary characters.

The definitive test would then be to make sure that JNI code is
able to properly deal with such Unicode character.

But doing that is out of scope here.