Make the libcurl dependency optional
Rationale: the way in which we use MK on Windows, Linux, and macOS
when compiling OONI probe is such that the iplookup is already
performed by the probe-engine. So, we can avoid cross-compiling
and linking with libcurl. This reduces the complexity of building
MK and also reduces the overall binary size.