Update to http-parser v2.9.4
While this is nice to have, it does not seem that we need to bless a new
MK release just because of this change. The version of http-parser that we
previously vendored is v2.9.3. Here's what changed since then:

https://github.com/nodejs/http-parser/compare/v2.9.3...v2.9.4

The only significant commit appears to be "Fix ABI breakage", i.e.,
https://github.com/nodejs/http-parser/commit/714cbb2dfbea9a521960b82b6239a991a64e3b9a.

We do not care about this change because we only use http-parser
internally and therefore the ABI does not matter to us.